Use a mixin module instead of inheritance
Prior to this commit, the module used a Puppet class to instantiate a
base provider for the other providers to inherit.  However, this caused
issues when using the 'require' metaparameter in manifests, because
Puppet may not have autoloaded the base type yet.  It was also overly
complicated.

This commit moves this functionality to a shared module which uses
constants and instance methods, which is a simpler design.  This removes
the need for the Puppet wrapper class, and the user does not need to
require anything to use the types.