Use token file on disk to fetch tokens
This commit changes the retrieve_token() function to use an admin token
on disk.  This allows it to be used as a deferred function where it will
be run on agents instead of the primary server.

Also fixes an issue where Dir.home would throw an error when running via
a service.